refactor: adding classes Group and Condition to handle data previously stored as objects, also moving model classes into model folder
