+++ README.md
@@ -13,3 +13,8 @@ A console app to call Virtuous, Query Contact information and write it to a csv 
   - Push a PR of the refactor with documentation on why you chose to change something
 
 You can find the API key for the exercise [here](https://keepersecurity.com/vault/share/#ea5rY4otBdLegGVuBPQIcc0ugJWPxHBIXlVQjlLAc2E)
+
+
+### Refactoring Updates
+ - I switched the Group and Condition logic to be strongly typed as that is a core feature of C# and I want the benefits of it when I am querying an external api.
+ - I also created a Models folder to group all POCOS (plain old C# / Clr / class objects), this is not for functionality but rather to ensure the codebase remains readable as it expands.
