+++ Sync/Models/AbbreviatedContact.cs
@@ -1,10 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Sync
+﻿
+namespace Sync.Models
 {
     public class AbbreviatedContact
     {
