+++ Sync/Models/Condition.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Generic;
+
+namespace Sync.Models
+{
+    public class Condition
+    {
+        public string Parameter { get; set; }
+        public string Operator { get; set; }
+        public string Value { get; set; }
+        public string SecondaryValue { get; set; }
+        public List<string> Values { get; set; }
+    }
+}
