+++ Sync/Models/Group.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace Sync.Models
+{
+    public class Group
+    {
+        public List<Condition> Conditions { get; set; }
+    }
+}
