+++ Sync/Models/PagedResult.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace Sync.Models
+{
+    public class PagedResult<T>
+    {
+        public int Total { get; set; }
+
+        public List<T> List { get; set; }
+    }
+}
