+++ Sync/Sync.csproj
@@ -84,11 +84,13 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AbbreviatedContact.cs" />
+    <Compile Include="Models\AbbreviatedContact.cs" />
     <Compile Include="Configuration.cs" />
-    <Compile Include="ContactQueryRequest.cs" />
+    <Compile Include="Models\Condition.cs" />
+    <Compile Include="Models\ContactQueryRequest.cs" />
     <Compile Include="DataAccessRepository.cs" />
-    <Compile Include="PagedResult.cs" />
+    <Compile Include="Models\Group.cs" />
+    <Compile Include="Models\PagedResult.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="VirtuousService.cs" />
